test: update e2e test for untp playground with error message
