+++ e2e/cypress/e2e/untp-playground/jsonld-expansion-validation.cy.ts
@@ -1,4 +1,4 @@
-import { VCDM_CONTEXT_URLS, allowedContextValue } from '../../../../packages/untp-playground/constants';
+import { allowedContextValue } from '../../../../packages/untp-playground/constants';
 
 describe('JSON-LD Expansion and Validation', () => {
   beforeEach(() => {
@@ -92,47 +92,5 @@ describe('JSON-LD Expansion and Validation', () => {
       'Properties "credentialSubject/invalid" are defined in the credential but missing from the context.',
       'Update the value(s) to the correct one(s) or remove the field(s).',
     ]);
-  });
-});
-
-describe('Display Error Messages', () => {
-  beforeEach(() => {
-    cy.visit('http://localhost:4000');
-  });
-
-  it('should show detail popup if upload file is not an object', () => {
-    cy.uploadCredential([]);
-    cy.contains('Fix validation error').click();
-    cy.contains('Expected type: object').should('be.visible');
-    cy.contains('Receive value:').should('be.visible');
-    cy.contains('[]').should('be.visible');
-    cy.contains('Example:').should('be.visible');
-    cy.contains(JSON.stringify(allowedContextValue, null, 2)).should('be.visible');
-  });
-
-  it('should view detail of error when `UNTP Schema Validation` failed', () => {
-    const invalidContextCredential = {
-      '@context': ['https://www.w3.org/ns/credentials/v2'],
-      type: ['VerifiableCredential', 'DigitalProductPassport'],
-      issuer: {
-        id: 'did:example:123',
-        name: 'dev',
-      },
-      credentialSubject: {
-        name: 'John Doe',
-        id: 'did:example:123',
-        type: ['Product'],
-      },
-    };
-
-    cy.uploadCredential(invalidContextCredential);
-    cy.expandGroup();
-    cy.checkValidationStatus('UNTP Schema Validation', 'failure');
-
-    cy.openErrorDetailsByStepName('UNTP Schema Validation');
-    cy.contains('Fix validation error').click();
-    cy.contains('Unable to access the schema from the constructed URI at test.uncefact.org').should('be.visible');
-    cy.contains('Example:').should('be.visible');
-    cy.contains(JSON.stringify(allowedContextValue, null, 2)).should('be.visible');
   });
 });
